dom/payments/test/test_showPayment.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1345366
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1345366</title>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript">

  "use strict";
  SimpleTest.waitForExplicitFinish();

  var gUrl = SimpleTest.getTestFileURL('ShowPaymentChromeScript.js');
  var gScript = SpecialPowers.loadChromeScript(gUrl);

  function testFailHandler(message) {
    ok(false, message);
  }
  function testPassHandler(message) {
    ok(true, message);
  }
  gScript.addMessageListener("test-fail", testFailHandler);
  gScript.addMessageListener("test-pass", testPassHandler);

  // testing data declaration
  // default parameters for PaymentRequest construction
  const defaultMethods = [{
    supportedMethods: "basic-card",
    data: {
      supportedNetworks: ['unionpay', 'visa', 'mastercard', 'amex', 'discover',
                          'diners', 'jcb', 'mir',
      ],
      supportedTypes: ['prepaid', 'debit', 'credit'],
    },
  }, {
    supportedMethods: "testing-payment-method",
  }];

  const defaultTotal = {
    label: "Total",
    amount: {
      currency: "USD",
      value: "1.00",
    },
  }

  const defaultDetails = {
    id: "test payment",
    total: defaultTotal,
    shippingOptions: [
      {
        id: "NormalShipping",
        label: "NormalShipping",
        amount: {
          currency: "USD",
          value: "10.00"
        },
        selected: false,
      },
      {
        id: "FastShipping",
        label: "FastShipping",
        amount: {
          currency: "USD",
          value: "30.00"
        },
        selected: false,
      },
    ],
  };

  const defaultOptions = {
    requestPayerName: true,
    requestPayerEmail: false,
    reqeustPayerPhone: false,
    requestShipping: true,
    shippingType: "shipping"
  };

  // testing data for PaymentRequestUpdateEvent.updateWith()
  const updatedShippingOptionsDetails = {
    total: defaultTotal,
    shippingOptions: [
      {
        id: "NormalShipping",
        label: "NormalShipping",
        amount: {
          currency: "USD",
          value: "10.00"
        },
        selected: false,
      },
      {
        id: "FastShipping",
        label: "FastShipping",
        amount: {
          currency: "USD",
          value: "30.00"
        },
        selected: true,
      },
    ],
  };

  const updatedErrorDetails = {
    total: defaultTotal,
    error: "Update with Error",
  };

  // Promise function for PaymentRequestUpdateEvent.updateWith()
  function updateWithPromise(detailsUpdate) {
    return new Promise((resolve, reject) => {
      if (detailsUpdate) {
        resolve(detailsUpdate);
      } else {
        reject();
      }
    });
  }

  // testing data for PaymentRequest.show() with Non-supported methods
  const nonSupportedMethods = [{
    supportedMethods: "nonsupported-method",
  }];


  // checking functions
  function checkAddress(address, fromEvent) {
    is(address.country, "USA", "address.country should be 'USA'.");
    is(address.region, "CA", "address.region should be 'CA'.");
    is(address.city, "San Bruno", "address.city should be 'San Bruno'.");
    is(address.dependentLocality, "Test locality",
       "address.dependentLocality should be 'Test locality'.");
    is(address.postalCode, "94066", "address.postalCode should be '94066'.");
    is(address.sortingCode, "123456", "address.sortingCode should be '123456'.");
    if (fromEvent) {
      is(address.addressLine.length, 0, "address.addressLine.length should be 0 from event.");
      is(address.organization, "", "address.organization should be empty from event.");
      is(address.recipient, "", "address.recipient should be empty from event.");
      is(address.phone, "", "address.phone should be empty from event.");
    } else {
      is(address.addressLine.length, 1, "address.addressLine.length should be 1 from promise.");
      is(address.addressLine[0], "Easton Ave", "address.addressLine[0] should be 'Easton Ave' from promise.");
      is(address.organization, "Testing Org", "address.organization should be 'Testing Org' from promise.");
      is(address.recipient, "Bill A. Pacheco", "address.recipient should be 'Bill A. Pacheco' from promise.");
      is(address.phone, "+1-434-441-3879", "address.phone should be '+1-434-441-3879' from promise.");
    }
  }

  function checkResponse(response) {
    is(response.requestId, "test payment", "response.requestId should be 'test payment'.");
    is(response.methodName, "testing-payment-method", "response.methodName should be 'testing-payment-method'.");
    is(response.details.paymentToken, "6880281f-0df3-4b8e-916f-66575e2457c1", "response.details.paymentToken should be '6880281f-0df3-4b8e-916f-66575e2457c1'.");
    checkAddress(response.shippingAddress, false/*fromEvent*/);
    is(response.shippingOption, "FastShipping", "response.shippingOption should be 'FastShipping'.");
    is(response.payerName, "Bill A. Pacheco", "response.payerName should be 'Bill A. Pacheco'.");
    ok(!response.payerEmail, "response.payerEmail should be empty");
    ok(!response.payerPhone, "response.payerPhone should be empty");
  }

  // testing functions
  function testShowNormalFlow() {
    gScript.sendAsyncMessage("set-normal-ui-service");
    return new Promise((resolve, reject) => {
      const payRequest = new PaymentRequest(defaultMethods, defaultDetails, defaultOptions);
      payRequest.addEventListener("shippingaddresschange", event => {
        checkAddress(payRequest.shippingAddress, true/*fromEvent*/);
        event.updateWith(updateWithPromise(defaultDetails));
      });
      payRequest.addEventListener("shippingoptionchange", event => {
        event.updateWith(updateWithPromise(updatedShippingOptionsDetails));
      });
      const handler = SpecialPowers.getDOMWindowUtils(window).setHandlingUserInput(true);
      payRequest.show().then(response => {
        checkResponse(response, false);
	checkAddress(payRequest.shippingAddress, false);
        response.complete().then(() =>{
          resolve();
        }).catch(e => {
          ok(false, "Unexpected error: " + e.name);
          resolve();
        });
      }).catch( e => {
        ok(false, "Unexpected error: " + e.name);
        resolve();
      }).finally(handler.destruct);
    });
  }

  // testing show with nonsupported methods
  function testCannotMakePaymentShow() {
    gScript.sendAsyncMessage("set-simple-ui-service");
    return new Promise((resolve, reject) => {
      const payRequest = new PaymentRequest(nonSupportedMethods, defaultDetails);
      const handler = SpecialPowers.getDOMWindowUtils(window).setHandlingUserInput(true);
      payRequest.canMakePayment().then(result => {
        ok(!result, "canMakePayment() should return false, but got " + result + ".");
        payRequest.show().then( () => {
          ok(false, "Should be rejected with 'NotSupportedError', but got resolved");
          resolve();
        }).catch( e => {
          is(e.name, "NotSupportedError", "Should be rejected with 'NotSupportedError', but got " + e.name + ".");
          resolve();
        });
      }).finally(handler.destruct);
    });
  }

  // testing show rejected by user
  function testRejectShow() {
    gScript.sendAsyncMessage("set-reject-ui-service");
    return new Promise((resolve, reject) => {
      const payRequest = new PaymentRequest(defaultMethods, defaultDetails, defaultOptions);
      const handler = SpecialPowers.getDOMWindowUtils(window).setHandlingUserInput(true);
      payRequest.show().then((result) => {
        ok(false, "Should be rejected with 'AbortError', but got resolved");
        resolve();
      }, (result) => {
        is(result.name, "AbortError", "Should be rejected with 'AbortError', but got " + result.name + ".");
        resolve();
      }).catch(e => {
        ok(false, "Unexpected error: " + e.name);
        resolve();
      }).finally(handler.destruct);
    });
  }

  // testing PaymentResponse.complete() with specified result
  function testCompleteStatus(result) {
    gScript.sendAsyncMessage("set-simple-ui-service");
    if (result) {
      gScript.sendAsyncMessage("set-complete-status-"+result);
    } else {
      gScript.sendAsyncMessage("set-complete-status-unknown");
    }
    return new Promise((resolve, reject) => {
      const payRequest = new PaymentRequest(defaultMethods, defaultDetails, defaultOptions);
      const handler = SpecialPowers.getDOMWindowUtils(window).setHandlingUserInput(true);
      payRequest.show().then(response => {
        response.complete(result).then(() => {
          resolve();
        }).catch(e => {
          ok(false, "Unexpected error: " + e.name);
          resolve();
        });
      }).catch( e => {
        ok(false, "Unexpected error: " + e.name);
        resolve();
      }).finally(handler.destruct);
    });
  }

  function testCompleteFail() {
    return testCompleteStatus("fail");
  }

  function testCompleteSuccess() {
    return testCompleteStatus("success");
  }

  function testCompleteUnknown() {
    return testCompleteStatus("unknown");
  }

  function testCompleteEmpty() {
    return testCompleteStatus();
  }

  // testing PaymentRequestUpdateEvent.updateWith with specified details and error
  function testUpdateWith(detailsUpdate, expectedError) {
    if (expectedError) {
      gScript.sendAsyncMessage("set-update-with-error-ui-service");
    } else {
      gScript.sendAsyncMessage("set-update-with-ui-service");
    }
    return new Promise((resolve, reject) => {
      const payRequest = new PaymentRequest(defaultMethods, defaultDetails, defaultOptions);
      payRequest.addEventListener("shippingaddresschange", event => {
        event.updateWith(updateWithPromise(detailsUpdate));
      });
      payRequest.addEventListener("shippingoptionchange", event => {
        event.updateWith(updateWithPromise(detailsUpdate));
      });
      const handler = SpecialPowers.getDOMWindowUtils(window).setHandlingUserInput(true);
      payRequest.show().then(response => {
        if (expectedError) {
          ok(false, "Should be rejected with " + expectedError + ", but got resolved");
	  resolve();
	} else {
	  response.complete("success").then(() => {
	    resolve();
	  })
	}
      }, response => {
        if (expectedError) {
          is(response.name, expectedError,
	     "Should be rejected with " + expectedError + ", but got " + response.name);
	} else {
	  ok(false, "Unexpected error: " + response.name);
	}
	resolve();
      }).catch(e => {
        ok(false, "Unexpected error: " + e.name);
	resolve();
      }).finally(handler.destruct);
    });
  }

  function testUpdateWithReject() {
    return testUpdateWith(null, "AbortError");
  }

  function testUpdateWithValidDetails() {
    return testUpdateWith(updatedShippingOptionsDetails, null);
  }

  function testUpdateWithInvalidDetails() {
    return testUpdateWith({total: "invalid details"}, "TypeError");
  }

  function testUpdateWithError() {
    return testUpdateWith(updatedErrorDetails, "AbortError");
  }

  // testing show with detailsUpdate promise
  function testShowWithDetailsPromise(detailsUpdate, expectedError) {
    if (expectedError) {
      gScript.sendAsyncMessage("set-reject-ui-service");
    } else {
      gScript.sendAsyncMessage("set-simple-ui-service");
    }
    return new Promise((resolve, reject) => {
      const payRequest = new PaymentRequest(defaultMethods, defaultDetails, defaultOptions);
      ok(!payRequest.shippingOption, "payRequest.shippingOption should be null.");
      const handler = SpecialPowers.getDOMWindowUtils(window).setHandlingUserInput(true);
      payRequest.show(updateWithPromise(detailsUpdate)).then(response => {
        if (expectedError) {
          ok(false, "Should be rejected with " + expectedError + ", but got resolved");
	  resolve();
	} else {
          ok(response.shippingOption, "response.shippingOption should not be null.");
	  response.complete().then(() => {
	    resolve();
	  })
	}
      }, response => {
        if (expectedError) {
          is(response.name, expectedError,
	     "Should be rejected with " + expectedError + ", but got " + response.name);
	} else {
	  ok(false, "Unexpected error: " + response.name);
	}
	resolve();
      }).catch(e => {
        ok(false, "Unexpected error: " + e.name);
	resolve();
      }).finally(handler.destruct);
    });
  }
  function testShowWithValidPromise() {
    return testShowWithDetailsPromise(updatedShippingOptionsDetails, null);
  }

  function testShowWithRejectedPromise() {
    return testShowWithDetailsPromise(null, "AbortError");
  }

  function testShowWithInvalidPromise() {
    return testShowWithDetailsPromise({total: "invalid details"}, "TypeError");
  }

  function testShowWithErrorPromise() {
    return testShowWithDetailsPromise(updatedErrorDetails, "AbortError");
  }

  function testShowWithPromiseResolvedByRejectedPromise() {
    gScript.sendAsyncMessage("set-reject-ui-service");
    return new Promise((resolve, reject)=> {
      const request = new PaymentRequest(defaultMethods, defaultDetails, defaultOptions);
      const handler = SpecialPowers.getDOMWindowUtils(window).setHandlingUserInput(true);
      let rejectPromise = Promise.reject(new TypeError());
      let detailsUpdatePromise = Promise.resolve(rejectPromise);
      request.show(detailsUpdatePromise).then(response => {
        ok(false, "should be rejected with 'AbortError', but got resolved.");
	resolve();
      }, response => {
        is(response.name, "AbortError", "Exepcted 'AbortError', but got " + response.name + ".");
	resolve();
      }).catch(error => {
        ok(false, "Unexpected error: " + error.name + ".");
      }).finally(handler.destruct);
    });
  }

  // testing show response initialization in chrome process
  function testShowResponseInit() {
    return new Promise((resolve, reject) => {
      gScript.addMessageListener("test-show-response-init-complete",
                                 function showResponseInitCompleteHandler() {
        gScript.removeMessageListener("test-show-response-init-complete",
                                      showResponseInitCompleteHandler);
        resolve();
      });
      gScript.sendAsyncMessage("test-show-response-init");
    });
  }

  // testing show that is not triggered by user.
  function testShowNotTriggeredByUser() {
    gScript.sendAsyncMessage("set-simple-ui-service");
    return new Promise((resolve, reject) => {
      const payRequest = new PaymentRequest(defaultMethods, defaultDetails);
      payRequest.show().then(() => {
        ok(false, "Expected 'SecurityError', but got resolved")
        resolve();
      }).catch((err) => {
        is(err.name, "SecurityError", `Should be rejected with 'SecurityError', but got ${err.name}`);
        resolve();
      });
    });
  }

  // teardown function
  function teardown() {
    gScript.addMessageListener("teardown-complete", function teardownCompleteHandler() {
      gScript.removeMessageListener("teardown-complete", teardownCompleteHandler);
      gScript.removeMessageListener("test-fail", testFailHandler);
      gScript.removeMessageListener("test-pass", testPassHandler);
      gScript.destroy();
      SimpleTest.finish();
    });
    gScript.sendAsyncMessage("teardown");
  }

  // test main body
  function runTests() {
    testCannotMakePaymentShow()
    .then(testRejectShow)
    .then(testShowNormalFlow)
    .then(testCompleteSuccess)
    .then(testCompleteFail)
    .then(testCompleteUnknown)
    .then(testCompleteEmpty)
    .then(testUpdateWithReject)
    .then(testUpdateWithValidDetails)
    .then(testUpdateWithInvalidDetails)
    .then(testUpdateWithError)
    .then(testShowWithValidPromise)
    .then(testShowWithInvalidPromise)
    .then(testShowWithRejectedPromise)
    .then(testShowWithErrorPromise)
    .then(testShowWithPromiseResolvedByRejectedPromise)
    .then(testShowResponseInit)
    .then(testShowNotTriggeredByUser)
    .then(teardown)
    .catch( e => {
      ok(false, "Unexpected error: " + e.name);
      SimpleTest.finish();
    });
  }

  window.addEventListener('load', function() {
    SpecialPowers.pushPrefEnv({
      'set': [
        ['dom.payments.request.enabled', true],
      ]
    }, runTests);
  });

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1345366">Mozilla Bug 1345366</a>
</body>
</html>
